Add apply transformation method to Vec4

+++ JTX/core/shader.hpp
@@ -13,7 +13,7 @@ namespace JTX::Core {
         virtual ~Shader() = default;
 
         virtual void vertex(const float *vertex, const float *normal, float* out) = 0;
-        virtual void fragment(const float *bary, const float *screen, const float *normal, float *color) = 0;
+        virtual void fragment(const float *bary, const int *screen, const float *normal, float *color) = 0;
         virtual void bind(const UniformBuffer& ub) = 0;
     };
 
@@ -23,7 +23,7 @@ namespace JTX::Core {
         //       potential switch to passing Vec3/Vec4 in the future
         void vertex(const float *vertex, const float *normal, float* out) override;
 
-        void fragment(const float *bary, const float *screen, const float *normal, float *color) override;
+        void fragment(const float *bary, const int *screen, const float *normal, float *color) override;
 
         void bind(const UniformBuffer& ub) override { this->ub_ = &ub; }
